[router] Handle initial tap link state.
When an interface is tapped, the tap inherits the link state of the
underlying device.

Also, the hw vector was being resized in the time between getting
the interface and when the hw_address was referenced, leading to
a segmentation fault. Resolve the issue by saving the mac
address contents on the stack before passing to other functions.

Change-Id: I4b5b31e438159a83ddfea808882503775b1fcd1a
Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>